+++ packages/wordpress/CHANGELOG.md
@@ -4,7 +4,7 @@
 - Added a new feature to cleanup inline styles from the block elements on the front end.
 
 ### Bug Fixes
-- Fixed a bug in the style engine optimization affecting the Render content of post and block.
+- Fixed a bug in the style engine optimization affecting the Render content of post and blocks.
 
 ## 1.0.1 (2024-12-21)