+++ packages/wordpress/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 1.1.0 (2025-03-15)
+
 ### New Features
 - Added a new feature to cleanup inline styles from the block elements on the front end.